Refactor display of space names

+++ CiRCLES/Views/More/MoreDatabaseAdministratiion.swift
@@ -16,9 +16,12 @@ struct MoreDatabaseAdministratiion: View {
     @Environment(AuthManager.self) var authManager
     @Environment(DatabaseManager.self) var database
 
+    @AppStorage(wrappedValue: false, "More.DBAdmin.SkipDownload") var willSkipDownload: Bool
+
     var body: some View {
         List {
             Section {
+                Toggle("More.DBAdmin.SkipDownload", isOn: $willSkipDownload)
                 Button("More.DBAdmin.RepairData", role: .destructive) {
                     withAnimation(.snappy.speed(2.0)) {
                         database.isBusy = true
@@ -26,18 +29,20 @@ struct MoreDatabaseAdministratiion: View {
                         UIApplication.shared.isIdleTimerDisabled = true
                         database.progressTextKey = ""
                         Task {
-                            if let token = authManager.token,
-                               let eventData = await WebCatalog.events(authToken: token),
-                               let latestEvent = eventData.list.first(where: {$0.id == eventData.latestEventID}) {
-                                database.delete()
-                                await MainActor.run {
-                                    database.progressTextKey = "Shared.LoadingText.DownloadTextDatabase"
+                            if !willSkipDownload {
+                                if let token = authManager.token,
+                                   let eventData = await WebCatalog.events(authToken: token),
+                                   let latestEvent = eventData.list.first(where: {$0.id == eventData.latestEventID}) {
+                                    database.delete()
+                                    await MainActor.run {
+                                        database.progressTextKey = "Shared.LoadingText.DownloadTextDatabase"
+                                    }
+                                    await database.downloadTextDatabase(for: latestEvent, authToken: token)
+                                    await MainActor.run {
+                                        database.progressTextKey = "Shared.LoadingText.DownloadImageDatabase"
+                                    }
+                                    await database.downloadImageDatabase(for: latestEvent, authToken: token)
                                 }
-                                await database.downloadTextDatabase(for: latestEvent, authToken: token)
-                                await MainActor.run {
-                                    database.progressTextKey = "Shared.LoadingText.DownloadImageDatabase"
-                                }
-                                await database.downloadImageDatabase(for: latestEvent, authToken: token)
                             }
                             if let textDatabaseURL = database.textDatabaseURL {
                                 do {
@@ -66,11 +71,12 @@ struct MoreDatabaseAdministratiion: View {
                                 } catch {
                                     debugPrint(error.localizedDescription)
                                 }
-                            }
-                            await MainActor.run {
-                                withAnimation(.snappy.speed(2.0)) {
-                                    database.isBusy = false
-                                    UIApplication.shared.isIdleTimerDisabled = false
+                            } else {
+                                await MainActor.run {
+                                    withAnimation(.snappy.speed(2.0)) {
+                                        database.isBusy = false
+                                        UIApplication.shared.isIdleTimerDisabled = false
+                                    }
                                 }
                             }
                         }
